Cleanup cassandra resources in tests explicitly
To ensure tests only attempt to delete present resources (not resources
about to be created by new tests).
We observed some test failures where the cassandra resource created by
the test was no found. Inspecting the diagnosis output showed that the
operator received a request to delete the cluster and then proceeded
accordingly.
As we perform a cleanup operation to remove all cassandra resources in between test suite,
the only plausible explanation is that the previous test bulk delete request
was somewhat deferred (perhaps the API server was overloaded at the time).

Test output showing cassandra resource not found
```
[1]       <*errors.StatusError | 0xc0003b6510>: {
[1]           ErrStatus: {
[1]               TypeMeta: {Kind: "", APIVersion: ""},
[1]               ListMeta: {SelfLink: "", ResourceVersion: "",
Continue: ""},
[1]               Status: "Failure",
[1]               Message: "cassandras.core.sky.uk \"mycluster-o5rct\"
not found",
[1]               Reason: "NotFound",
[1]               Details: {
[1]                   Name: "mycluster-o5rct",
[1]                   Group: "core.sky.uk",
[1]                   Kind: "cassandras",
[1]                   UID: "",
[1]                   Causes: nil,
[1]                   RetryAfterSeconds: 0,
[1]               },
[1]               Code: 404,
[1]           },
[1]       }
[1]       cassandras.core.sky.uk "mycluster-o5rct" not found
[1]   not to have occurred
```