Fix panic when comparing snapshots (#120)
* WIP - trying to prove concurrency issue

* Add diagnosis for statefulset and cassandra

Also print the server's representation of the cassandra specs on update

* pod start timeout applies regardless of mock mode

This is because the slowest point in the pod start-up sequence is
usually the volume attachment, which can take several minutes even when
using a local provisioner.

* Better debug output for struct

Use %+v to output fields and spew to print embedded struct
Debug the event type received, but not the fields in the event Data
as this can contain configmap and potentially sensitive info.

* Fix panic when comparing snaphot properties

As some properties can be nillable.
Also move helper function to helpers file and add tests for it

* Remove tests that did not prove the issue

* Delete cassandra dependents automatically

Change the delete propagation policy to Background, as using Foreground
or Orphan results in an additional cluster update event before the
cluster delete event. Before this PR, this would mean the cluster
cleanup was triggering a cluster update event which would in some rare
cases cause a panic when comparing the Snapshot properties and so cause
other tests (when running in parallel) to fail as their update was lost
during the operator restart.

Using a Background policy also mimics better the behaviour observed when
deleting a `cassandra` resource using `kubectl`. In fact using the
`--cascade=false` option also causes an additional cluster update event
when deleting a `cassandra` resource

* Wait for pods to be deleted

As since the cassandra resource is deleted in background mode
we need to wait long enough for the GC to do its job

* Use logrus fields to standardised logging a bit better

* Fix typo

* Update cassandra-operator/pkg/apis/cassandra/v1alpha1/helpers/cassandra_test.go

Co-Authored-By: alangibson01 <alan.gibson@sky.uk>